(PWGCF): QVectors changed to separate headers, HistManager moved
