+++ PWGCF/Flow/Core/CMakeLists.txt
@@ -10,12 +10,10 @@
 # or submit itself to any jurisdiction.
 
 o2physics_add_library(PWGCFFlowCore
-                        SOURCES FlowJHistManager.cxx
-                          FlowJSPCAnalysis.cxx
-                            PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore)
+                        SOURCES FlowJSPCAnalysis.cxx
+                          PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore)
 
 o2physics_target_root_dictionary(PWGCFFlowCore
-                                     HEADERS FlowJHistManager.h
-                                       FlowJSPCAnalysis.h
-                                         FlowJSPCObservables.h
+                                     HEADERS FlowJSPCAnalysis.h
+                                        FlowJSPCObservables.h
                                            LINKDEF PWGCFFlowLinkDef.h)
